fix account: match C++ retry-after parsing, credentials load type-strictness, atomic write close()/temp-name, model cache fsync, and silent browser-opener exec failure
Findings 2, 3, 4, 5, 6, 7 from /tmp/wally-rust-migration/fixes/account.json.

- console.rs: parse Retry-After as i32 (not i64-then-clamp) so an
  overflowing value reports -1 like std::from_chars, instead of clamping
  to a day and stalling login retry/poll cadence.
- credentials.rs: load() now errors on a present, wrong-JSON-typed field
  (e.g. access_token as a number) instead of silently defaulting it,
  matching nlohmann's get<T>() throwing on a type mismatch.
- credentials.rs: write_document (non-Windows) explicitly closes the
  temp file's fd and fails the write (leaving the old file in place) if
  close() fails, instead of discarding that error via File::drop.
- credentials.rs: the temp file is now a visible sibling
  (credentials.json.tmp.XXXXXX), matching C++'s mkstemp template,
  instead of a hidden, unrelated basename.
- model_cache.rs: write_cache no longer fsyncs (C++ never does), so a
  filesystem where fsync fails but buffered writes succeed no longer
  silently leaves the cache stale.
- cmd_account.rs: open_browser now forks/execvp's directly instead of
  going through std::process::Command, so a missing opener binary stays
  silent like C++ (only a failed fork() prints the fallback line).
- Adds tests/test_wally_fix_account.rs plus inline #[cfg(test)] units in
  credentials.rs and cmd_account.rs.

+++ src/account/console.rs
@@ -42,8 +42,11 @@ impl HttpResponse {
         if value.is_empty() || !value.bytes().all(|b| b.is_ascii_digit()) {
             return -1;
         }
-        match value.parse::<i64>() {
-            Ok(seconds) => std::cmp::min(seconds, 86400) as i32,
+        // Parse into a 32-bit width (matches C++'s `int` via std::from_chars):
+        // a digit string that overflows i32 (e.g. an accidental millisecond
+        // epoch) must report "no valid Retry-After", not clamp to the ceiling.
+        match value.parse::<i32>() {
+            Ok(seconds) => std::cmp::min(seconds, 86400),
             Err(_) => -1,
         }
     }
